IS: Refactor, delete ToolbarWrapper
